requirements.txt
google-generativeai
Pillow 
django-environ